for-cicd/requirements.txt
requests
PyYaml
ansible
